scripts/requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.26.0
tqdm>=4.62.0 